+++ src/input_condition/hold_and_release.rs
@@ -2,8 +2,8 @@ use bevy::prelude::*;
 
 use super::{DEFAULT_ACTUATION, InputCondition, condition_timer::ConditionTimer};
 use crate::{
+    action_map::{ActionMap, ActionState},
     action_value::ActionValue,
-    actions::{ActionState, ActionsData},
 };
 
 /// Returns [`ActionState::Ongoing`] when input becomes actuated and [`ActionState::Fired`]
@@ -48,7 +48,7 @@ impl HoldAndRelease {
 impl InputCondition for HoldAndRelease {
     fn evaluate(
         &mut self,
-        _actions: &ActionsData,
+        _action_map: &ActionMap,
         time: &Time<Virtual>,
         value: ActionValue,
     ) -> ActionState {
@@ -77,12 +77,12 @@ mod tests {
     use core::time::Duration;
 
     use super::*;
-    use crate::actions::ActionsData;
+    use crate::action_map::ActionMap;
 
     #[test]
     fn hold_and_release() {
         let mut condition = HoldAndRelease::new(1.0);
-        let actions = ActionsData::default();
+        let actions = ActionMap::default();
         let mut time = Time::default();
 
         assert_eq!(